adding in stats functionality for centralizing data

+++ grails-app/domain/net/nosegrind/apiframework/Stat.groovy
@@ -2,7 +2,7 @@ package net.nosegrind.apiframework
 
 class Stat {
 
-	Integer user
+	Long user
 	Integer code
 	String uri
 	Long timestamp
